xfs_metadump: manpage fix regarding frozen fs
The xfs_metadump manpage states that metadump works
on a frozen filesystem; it does not.  In fact, there is
no way to detect a frozen filesystem, so we can't make it
work, either.

So just remove this from the manpage; unmounted or RO
mounted is what is enforced by xfs_metadump.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>